+++ CHANGELOG.md
@@ -3,3 +3,13 @@
 ## v0.1.2
 ### Added
 - Add packaging configurations. Can now be run with `sitegen`.
+
+## v0.1.3
+### Added
+- Added context variable for the current page URL.
+- Added fallback template for a page.
+- Added fallback template for RSS feed.
+- Added fallback template for sitemap.
+
+### Fixes
+- Fixed issue where blank page default was not populated.